카프카 기본동작원리(2.6)/app22/pyproject.toml
[project]
name = "app22"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.2",
    "kafka-python>=2.3.0",
    "pydantic-settings>=2.12.0",
]
